// src/types/generated/ArticleAlerte.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ArticleAlerte = { id: number, designation: string, stock: number, stock_alerte: number, categorie_nom: string | null, fournisseur_nom: string | null, fournisseur_id: number | null, prix_achat: number, suggestion_qte: number, };